restructure package into modules
This commit restructures the package into 5 different python modules and an IO subpackage.
