+++ molecool/__init__.py
@@ -4,7 +4,12 @@ A Python package for analyzing and visualizing molecular files. For MolSSI works
 """
 
 # Add imports here
-from .functions import *
+from .functions import canvas, zen
+from .measure import calculate_angle, calculate_distance
+from .molecule import build_bond_list
+from .visualize import draw_bond_histogram, draw_molecule
+
+import molecool.io
 
 # Handle versioneer
 from ._version import get_versions
