+++ molecool/io/__init__.py
@@ -1,0 +1,6 @@
+'''
+init for IO subpackage
+'''
+
+from .pdb import open_pdb
+from .xyz import open_xyz, write_xyz
